fix(mastodon): cast attachment id to int in Media module

+++ src/Module/Api/Mastodon/Media.php
@@ -94,7 +94,7 @@ class Media extends BaseApi
 		}
 
 		if (DI::mstdnAttachment()->isAttach($this->parameters['id']) && Attach::exists(['id' => substr((string) $this->parameters['id'], 7)])) {
-			$this->earlyJsonExit(DI::mstdnAttachment()->createFromAttach(substr((string) $this->parameters['id'], 7)));
+			$this->earlyJsonExit(DI::mstdnAttachment()->createFromAttach((int) substr((string) $this->parameters['id'], 7)));
 		}
 
 		$photo = Photo::selectFirst(['resource-id'], ['id' => $this->parameters['id'], 'uid' => $uid]);
@@ -144,7 +144,7 @@ class Media extends BaseApi
 		}
 
 		if (DI::mstdnAttachment()->isAttach($id) && Attach::exists(['id' => substr((string) $id, 7)])) {
-			$this->earlyJsonExit(DI::mstdnAttachment()->createFromAttach(substr((string) $id, 7)));
+			$this->earlyJsonExit(DI::mstdnAttachment()->createFromAttach((int) substr((string) $id, 7)));
 		}
 
 		$this->logAndJsonError(404, $this->errorFactory->RecordNotFound());
